fixed the build and couple of minor issues.
